backend_foodgram/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-colorfield==0.10.1
django-filter==23.2
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
et-xmlfile==1.1.0
filetype==1.2.0
flake8==6.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.1
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.5
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
xlrd==2.0.1
xlwt==1.3.0